+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+  * `TlsConfig::unversioned_rustls_crypto_provider()` (#931)
+  * Feature `rustls-no-provider` to compile without ring (#931)
   * Re-enable CONNECT proxy support (#932)
   * Body::content_length (#927)
   * Handle Authorization: Basic from URI (#923)